+++ dynomite/trybuild-tests/item-has-no-partition-key.rs
@@ -1,0 +1,8 @@
+use dynomite_derive::Item;
+
+#[derive(Item)]
+struct Foo {
+    key: String
+}
+
+fn main() {}